refactor / restructure
